+++ src/classes/attack.js
@@ -1,10 +1,15 @@
+import { attacksDB } from "../data/attacks.db";
+import { powerRanks } from "../data/ranks.db";
+
 class Attack{
   constructor(attackName){
     this.attackName = attackName;
-    this.maxCost = 8;
-    this.currCost = 8;
+    this.type = attacksDB[this.attackName].type;
+    this.maxCost = attacksDB[this.attackName].maxCost;
+    this.currCost = 6;
 
-    this.power = 4;
+    this.power = attacksDB[this.attackName].power;
+    this.targets = attacksDB[this.attackName].targets;
   }
 }