+++ src/classes/menu/dgmn-attack-menu.js
@@ -8,7 +8,9 @@ class DgmnAttackMenu {
     this.currentChoice = 0;
 
     this.textManagers = [
-      new TextManager([fetchImageCallback('fontsWhite')], 1,15, 5,2 )
+      new TextManager([fetchImageCallback('fontsWhite')], 1,15, 5,2 ),
+      new TextManager([fetchImageCallback('fontsWhite')], 1,15, 5,4 ),
+      new TextManager([fetchImageCallback('fontsWhite')], 1,15, 5,6 )
     ];
 
     this.fetchImage = imageName => { return fetchImageCallback(imageName) }
@@ -18,22 +20,25 @@ class DgmnAttackMenu {
     this.attackList = attackData;
   }
 
-  refreshList = (canvas, index) => {
-    let attack = this.attackList[index];
-    this.textManagers[0].instantPaint(canvas,attacksDB[attack.attackName].displayName);
+  refreshList = (canvas) => {
+    let loopCount = this.attackList.length > 6 ? 6 : this.attackList.length;
+    for(let i = 0; i < loopCount; i++){
+      let attack = this.attackList[i];
+      let targetCount = attack.targets === 'single' ? 'targetOne' : 'targetAll';
+      this.textManagers[i].instantPaint(canvas,attacksDB[attack.attackName].displayName);
+  
+      canvas.ctx.drawImage(this.fetchImage('costLabel'), 5 * (8 * config.screenSize), ( 3 + (2 * i) ) * (8 * config.screenSize), 16 * config.screenSize, 8 * config.screenSize);
+  
+      this.calculateCost(canvas,attack,i);
 
-    canvas.ctx.drawImage(this.fetchImage('costLabel'), 5 * (8 * config.screenSize), 3 * (8 * config.screenSize), 16 * config.screenSize, 8 * config.screenSize);
-
-    this.calculateCost(canvas,attack)
-
-    canvas.ctx.drawImage(this.fetchImage('fireTypeIcon'), 16 * (8 * config.screenSize), 3 * (8 * config.screenSize), 8 * config.screenSize, 8 * config.screenSize);
-    canvas.ctx.drawImage(this.fetchImage('pwrEIcon'), 17 * (8 * config.screenSize), 3 * (8 * config.screenSize), 8 * config.screenSize, 8 * config.screenSize);
-    canvas.ctx.drawImage(this.fetchImage('targetOne'), 18 * (8 * config.screenSize), 3 * (8 * config.screenSize), 8 * config.screenSize, 8 * config.screenSize);
-    canvas.ctx.drawImage(this.fetchImage('oneHitIcon'), 19 * (8 * config.screenSize), 3 * (8 * config.screenSize), 8 * config.screenSize, 8 * config.screenSize);
-
+      canvas.ctx.drawImage(this.fetchImage(`${attack.type}TypeIcon`), 16 * (8 * config.screenSize), ( 3 + (2 * i) ) * (8 * config.screenSize), 8 * config.screenSize, 8 * config.screenSize);
+      canvas.ctx.drawImage(this.fetchImage(`pwr${attack.power}Icon`), 17 * (8 * config.screenSize), ( 3 + (2 * i) ) * (8 * config.screenSize), 8 * config.screenSize, 8 * config.screenSize);
+      canvas.ctx.drawImage(this.fetchImage(targetCount), 18 * (8 * config.screenSize), ( 3 + (2 * i) ) * (8 * config.screenSize), 8 * config.screenSize, 8 * config.screenSize);
+      canvas.ctx.drawImage(this.fetchImage('oneHitIcon'), 19 * (8 * config.screenSize), ( 3 + (2 * i) ) * (8 * config.screenSize), 8 * config.screenSize, 8 * config.screenSize);
+    }
   }
 
-  calculateCost = (canvas, attack) => {
+  calculateCost = (canvas, attack, index) => {
     let blockCount = attack.maxCost / 4;
     let remCount = attack.currCost;
     for(let i = 0; i < blockCount; i++){
@@ -42,7 +47,7 @@ class DgmnAttackMenu {
       let final = 25 * (remCount - check);
           final = final >= 0 ? 0 : final;
           final = final < 0 ? -100 : final;
-      canvas.ctx.drawImage(this.fetchImage(`costMeter${100 + final}`), ( (7 + (i)) * (8 * config.screenSize) ), 3 * (8 * config.screenSize), (8 * config.screenSize), (8 * config.screenSize)  );
+      canvas.ctx.drawImage(this.fetchImage(`costMeter${100 + final}`), ( (7 + (i)) * (8 * config.screenSize) ), (3 + (index * 2) ) * (8 * config.screenSize), (8 * config.screenSize), (8 * config.screenSize)  );
       remCount -= 4;
     }
   }