+++ src/data/attacks.db.js
@@ -1,7 +1,34 @@
 export const attacksDB = {
+  bubbles: {
+    displayName: 'Bubbles',
+    type: 'none',
+    power: 'F',
+    maxCost: 32,
+    targets: 'single',
+    hits: 1
+  },
   babyFlame: {
     displayName: 'Baby Flame',
+    power: 'E',
+    maxCost: 8,
     type: 'fire',
-    targets: 'single'
+    targets: 'single',
+    hits: 1
+  },
+  megaFlame: {
+    displayName: 'Mega Flame',
+    power: 'D',
+    maxCost: 8,
+    type: 'fire',
+    targets: 'single',
+    hits: 1
+  },
+  petitFire: {
+    displayName: 'Petit Fire',
+    power: 'F',
+    maxCost: 8,
+    type: 'fire',
+    targets: 'all',
+    hits: 1
   }
 }