+++ src/data/ranks.db.js
@@ -1,0 +1,3 @@
+// Handles all of the actual values for Ranks (F - S)
+
+export const powerRanks = { F: 4, E: 5, D: 6, C: 7, B: 8, A: 9, S: 10 }