+++ CHANGELOG.md
@@ -1,14 +1,20 @@
 # DIGIMON : TOWER OF RUIN
 
-## CURRENT VERSION : 0.0.0-alpha.5
+## CURRENT VERSION : 0.0.0-alpha.6
 
 ### 0.0.0
 Setting up all of the basic structure for the system, until the game is in any way testable with other users
 
+### v0.0.0-alpha.6
+* Cleaned up the battle system
+* Added more than one possible attacker
+* Added Defend Option
+
 #### v0.0.0-alpha.5
 * Built Battle Menu
-  * Create Status Menus with HP and EN
-  * Create Attack Menus
+  * Created Status Menus with HP and EN
+  * Created Attack Menus
+  * Added stripped-down attack functionality
 * Built Text Sprite Sheet Functionality
 
 #### v0.0.0-alpha.4